refactor code and added close button
