+++ Basic-Car-Maintenance/Shared/Dashboard/Views/CSVEncoder.swift
@@ -7,55 +7,6 @@
 //
 
 import Foundation
-import SwiftUI
-
-struct CSVFile {
-    var events: [MaintenanceEvent]
-    
-    func csvData() -> String {
-        let table = CSVTable<MaintenanceEvent>(
-            columns: [
-                CSVColumn("Date") { $0.date.formatted() },
-                CSVColumn("Vehicle Name", \.title),
-                CSVColumn("Notes", \.notes)
-            ], 
-            configuration: CSVEncoderConfiguration(dateEncodingStrategy: .iso8601) 
-        )
-        return table.export(rows: events)
-    }
-    
-    func generateCSVFile(vehicle: String) -> URL? {
-        // Get the path to the Documents Directory
-        let fileManager = FileManager.default
-            guard let documentsDirectory = fileManager.urls(
-                for: .documentDirectory, in: .userDomainMask).first else {
-                print("Failed to locate the Documents Directory.")
-                return nil
-            }
-        
-        // Create the file URL
-        let fileName = "\(vehicle)-MaintenanceReport"
-        let fileURL = documentsDirectory.appendingPathComponent(fileName).appendingPathExtension("csv")
-        
-        do {
-                // Save the CSV content to the file
-            try csvData().write(to: fileURL, atomically: true, encoding: .utf8)
-                print("File saved to \(fileURL)")
-                return fileURL
-            } catch {
-                print("Failed to save CSV file: \(error.localizedDescription)")
-                return nil
-            }
-    }
-}
-
-extension CSVFile: Transferable {
-  static var transferRepresentation: some TransferRepresentation {
-    DataRepresentation(exportedContentType: .commaSeparatedText) { file in
-      Data(file.csvData().utf8)
-    }
-  }
-}
 
 internal extension BidirectionalCollection where Element == String {
     var commaDelimited: String { joined(separator: ",") }
